Testing the listener efficiency
